requirements.txt
Flask==2.0.1
PyJWT==2.8.0
Requests==2.31.0
